Make it easier to distinguish between slot and signal trampolines
We generate "foo_slot_trampoline" and "foo_signal_handler_trampoline"
names now.  The former is when C code wants to call the Rust
implementation of a slot (e.g. a virtual method implementation or a
signal's default handler).  The latter is the callback passed to
g_signal_connect(); it must then call a Rust closure with the signal
emission's arguments.