Parse "impl interface Foo for Bar" to implement an interface for a class
If we just used "impl Foo for MyClass", we can't know whether Foo
corresponds to a class whose methods we are overriding, or to an
interface that we want to implement.  So, with "impl interface" we
make this unambiguous.

This may change in the future if we have the Gir information at hand.